kloud: initialize logger inside attachsession to avoid panic
